modular structure, tests, and docs
