+++ appveyor.yml
@@ -18,20 +18,6 @@ environment:
   # Stable 32-bit MSVC
     - channel: stable
       target: i686-pc-windows-msvc
-  # Beta 64-bit MSVC
-    - channel: beta
-      target: x86_64-pc-windows-msvc
-  # Beta 32-bit MSVC
-    - channel: beta
-      target: i686-pc-windows-msvc
-  # Nightly 64-bit MSVC
-    - channel: nightly
-      target: x86_64-pc-windows-msvc
-      #cargoflags: --features "unstable"
-  # Nightly 32-bit MSVC
-    - channel: nightly
-      target: i686-pc-windows-msvc
-      #cargoflags: --features "unstable"
 
 ### GNU Toolchains ###
 
@@ -41,26 +27,6 @@ environment:
   # Stable 32-bit GNU
     - channel: stable
       target: i686-pc-windows-gnu
-  # Beta 64-bit GNU
-    - channel: beta
-      target: x86_64-pc-windows-gnu
-  # Beta 32-bit GNU
-    - channel: beta
-      target: i686-pc-windows-gnu
-  # Nightly 64-bit GNU
-    - channel: nightly
-      target: x86_64-pc-windows-gnu
-      #cargoflags: --features "unstable"
-  # Nightly 32-bit GNU
-    - channel: nightly
-      target: i686-pc-windows-gnu
-      #cargoflags: --features "unstable"
-
-### Allowed failures ###
-matrix:
-  allow_failures:
-    - channel: nightly
-    - channel: beta
 
 ## Install Script ##
 install:
@@ -74,6 +40,9 @@ install:
 build: false
 test_script:
   - cargo test --verbose %cargoflags%
+  - cargo run --example cross-threads
+  - cargo run --example logger-demo
+  - cargo run --example panic-handler-demo
 
 ## Notifications ##
 notifications:
